refactor: orderTable 이벤트리스너 제거
